+++ slow_website_alert.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 ###
 #
@@ -8,8 +8,8 @@
 #                   slower than a specific threshold, the script sends an alert.
 #          Author:  Elliot Jordan <elliot@elliotjordan.com>
 #         Created:  2014-12-02
-#   Last Modified:  2014-12-04
-#         Version:  1.0.1
+#   Last Modified:  2014-12-20
+#         Version:  1.0.2-beta
 #
 ###
 
@@ -26,7 +26,7 @@ URL=(
 
 # The name of the log file to which speed results will be saved. By default,
 # this is the name of this script with .log extension.
-LOG_FILE="$(basename $0 | sed 's/.sh/.log/g')"
+LOG_FILE="$(basename "$0" | sed 's/.sh/.log/g')"
 
 
 ################################ ALERT SETTINGS ################################
@@ -42,7 +42,7 @@ TEST_FREQ=300
 EMAIL_TO="you@pretendco.com, somebodyelse@pretendco.com"
 
 # The email notifications will be sent from this email address.
-EMAIL_FROM="$(echo $(whoami)@$(hostname))"
+EMAIL_FROM="$(whoami)@$(hostname)"
 
 # The path to sendmail on your server. Typically /usr/sbin/sendmail.
 sendmail="/usr/sbin/sendmail"
@@ -61,12 +61,30 @@ DEBUG_MODE=false
 ################################################################################
 
 
+################################## FUNCTIONS ###################################
+
+# Log functions
+APPNAME=$(basename "$0" | sed "s/\.sh$//")
+fn_log_info() {
+    echo "$(date) : $APPNAME : $1" >> "$LOG_FILE"
+}
+fn_log_debug() {
+    echo "$(date) : $APPNAME : [DEBUG] $1" >> "$LOG_FILE"
+}
+fn_log_warn() {
+    echo "$(date) : $APPNAME : [WARNING] $1" >> "$LOG_FILE"
+}
+fn_log_error() {
+    echo "$(date) : $APPNAME : [ERROR] $1" >> "$LOG_FILE"
+}
+
+
 ######################## VALIDATION AND ERROR CHECKING #########################
 
 # Let's make sure we aren't using default email alert settings.
 if [[ $EMAIL_TO == "you@pretendco.com, somebodyelse@pretendco.com" ]]; then
 
-    echo "Error: The email alert settings are still set to the default value. Please edit them to suit your environment." >&2
+    fn_log_error "The email alert settings are still set to the default value. Please edit them to suit your environment."
     exit 1001
 
 fi # End email alert settings validation.
@@ -74,14 +92,14 @@ fi # End email alert settings validation.
 # Let's make sure the sendmail path is correct.
 if [[ ! -x "$sendmail" ]]; then
 
-    echo "Error: The specified path to sendmail ($sendmail) appears to be incorrect. Trying to locate the correct path..." >&2
+    fn_log_warn "The specified path to sendmail ($sendmail) appears to be incorrect. Trying to locate the correct path..."
     sendmail_try2=$(which sendmail)
 
     if [[ $sendmail_try2 == '' || ! -x $sendmail_try2 ]]; then
-        echo "Error: Unable to locate the path to sendmail." >&2
+        fn_log_error "Unable to locate the path to sendmail."
         exit 1002
     else
-        echo "Located sendmail at $sendmail_try2. Please adjust the \"$(basename $0)\" script settings accordingly." >&2
+        fn_log_warn "Located sendmail at $sendmail_try2. Please adjust the $APPNAME script settings accordingly."
         sendmail="$sendmail_try2"
         # Fatal error avoided. No exit needed.
     fi
@@ -92,7 +110,7 @@ fi # End sendmail validation.
 if (( $(bc <<< "$TEST_FREQ < 5") == 1 )); then
     if [[ $DEBUG_MODE == false ]]; then
 
-        echo "Error: It's not recommended to run tests less than 5 seconds apart." >&2
+        fn_log_error "It's not recommended to run tests less than 5 seconds apart."
         exit 1003
 
     fi
@@ -105,7 +123,7 @@ if [[ $USE_TERMINAL_NOTIFIER == true ]]; then
 
     if [[ "$notifier" == "" || ! -x "$notifier" ]]; then
 
-        echo "Error: Terminal Notifier is not installed or not executable." >&2
+        fn_log_error "Terminal Notifier is not installed or not executable."
         exit 1004
 
     fi
@@ -117,7 +135,7 @@ if (( $(bc <<< "$MAX_TIME > 60") == 1 ||
       $(bc <<< "$MAX_TIME < 2") == 1 )); then
     if [[ $DEBUG_MODE == false ]]; then
 
-        echo "Error: It's recommended to set a MAX_TIME between 2 seconds and 60 seconds." >&2
+        fn_log_error "It's recommended to set a MAX_TIME between 2 seconds and 60 seconds."
         exit 1005
 
     fi
@@ -130,19 +148,22 @@ fi # End MAX_TIME validation.
 SITE_COUNT=${#URL[@]}
 
 # Let the user know how to stop the loop.
-echo "Website speed tests have started. To stop, press Control-C."
+fn_log_info "Website speed tests have started. To stop, press Control-C."
 
 # Begin looping.
 while [[ true ]]; do
 
     # Begin iterating through websites.
-    for (( i = 0; i < $SITE_COUNT; i++ )); do
+    for (( i = 0; i < SITE_COUNT; i++ )); do
 
-        TIME=$(curl -s -w %{time_total}\\n -o /dev/null "${URL[$i]}")
-        printf "\n$(date) : $(hostname) : ${URL[$i]} : $TIME seconds" >> "$LOG_FILE"
+        TIME1=$(curl -s -w "%{time_total}" -o /dev/null "${URL[$i]}")
+        TIME2=$(curl -s -w "%{time_total}" -o /dev/null "${URL[$i]}")
+        TIME3=$(curl -s -w "%{time_total}" -o /dev/null "${URL[$i]}")
+        AVG_TIME=$(bc <<< "scale=3; ( $TIME1 + $TIME2 + $TIME3 ) / 3")
+        fn_log_info "$(date) : $(hostname) : ${URL[$i]} : $AVG_TIME seconds"
 
-        if (( $(bc <<< "$TIME > $MAX_TIME") == 1 )); then
-            printf " (greater than $MAX_TIME)" >> "$LOG_FILE"
+        if (( $(bc <<< "$AVG_TIME > $MAX_TIME") == 1 )); then
+            fn_log_warn "Last result was greater than $MAX_TIME."
             
             # Construct an email.
             EMAIL_SUBJ="[${URL[$i]}] Website response greater than $MAX_TIME seconds"
@@ -150,26 +171,26 @@ while [[ true ]]; do
             EMAIL_MSG+="Latest log entry:\n$(tail -n 1 "$LOG_FILE")\n\n"
             EMAIL_MSG+="Thank you.\n\n"
             EMAIL_MSG+="\n\n"
-            EMAIL_MSG+="(This is an automated message sent by the \"$(basename $0)\" script running on $(hostname).)"
+            EMAIL_MSG+="(This is an automated message sent by the $APPNAME script running on $(hostname).)"
 
             THE_EMAIL="From: $EMAIL_FROM\nTo: $EMAIL_TO\nSubject: $EMAIL_SUBJ\n$EMAIL_MSG\n.\n"
 
             if [[ $DEBUG_MODE == true ]]; then
                 # Print the message, if in debug mode.
-                printf "\n$THE_EMAIL\n"
+                printf "\n%s\n" "$THE_EMAIL"
             elif [[ $DEBUG_MODE == false ]]; then
-                #statements
+                
                 # Send the message.
-                printf "$THE_EMAIL" | $sendmail "$EMAIL_TO"
+                printf "%s" "$THE_EMAIL" | $sendmail "$EMAIL_TO"
 
                 # Send notification to Terminal.
                 if [[ $USE_TERMINAL_NOTIFIER == true ]]; then
-                    $notifier -message "$TIME seconds" -title "Slow website alert" -subtitle "${URL[$i]}" -sound Sosumi -open "file://$(pwd)/$LOG_FILE" -group $(date +%s)
+                    $notifier -message "$AVG_TIME seconds" -title "Slow website alert at $(date +%H:%M)" -subtitle "${URL[$i]}" -sound Sosumi -open "file://$(pwd)/$LOG_FILE" -group "$(date +%s)"
                 fi
 
             else
 
-                echo "Error: DEBUG_MODE must be set to either \"true\" or \"false\"." >&2
+                fn_log_error "DEBUG_MODE must be set to either \"true\" or \"false\"."
                 exit 1011
 
             fi