+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-WP Plugin Counter Changelog
-===========================
+Slow Website Alert Changelog
+============================
+
+Version 1.0.2 - 2014-12-23
+
+- Response time is now the average of three `curl` tests.
+- Now includes timestamp in Terminal Notifier alerts.
+- Using functions for log output.
+- Minor changes based on shell script linter feedback.
 
 Version 1.0.1 - 2014-12-04